[WIP!!] dbld: remove local dir usage in rpm spec file in case of source files.
Some files are needed for RPM generating (service file, default config).
In the spec file they have the local dir as path.
The local dir will be the RPMBUILD directory.
As these files are not existing when rpm generation starts, the 'dbld/rpm'
script is copying them.

This patch changes the path of these files to use the ones
under packaging dir.