packaging/rhel: fix python2/python3 build dependency installs on centos-7
The `%{with python2}` and `%{with python3}` macros always evaluated to TRUE (on dbld centos-7 image),
even if --without-python2/3 is used and the %{with_python2} macros are undefined or set explicitly to FALSE.
This results that syslog-ng is always built with python3.

Using the macro %{with_python2/3} explicitly in if statements instead of the ${with} "function" solves the issue.

Signed-off-by: Gabor Nagy <gabor.nagy@oneidentity.com>