dbld: change signature of make command
All of our other commands in dbld have the following signature:
command-<image-name> <COMMAND_ARGS>=<...>

The make command always runs in the default image, and after the dash
we give the command arg == the make target.

This commit intends to change the make command's signature to match all
of our other commands.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>