Light: make pytest-linters work in build dir outside of git tree
`pre-commit` needs to run in a git source. It is used in a make target
so it gets executed from the build dir. The build dir is usually inside
the source dir for most of us, but dbld puts it to completely separate
dirs.

This commit makes sure that `pre-commit` is executed from the source dir.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>